Move spectral model into model.wsclean
