+++ HISTORY.rst
@@ -5,7 +5,7 @@ History
 0.1.3 (2018-03-28)
 ------------------
 * Gaussian Shape Parameter Implementation (:pr:`82`, :pr:`83`)
-* Spectral Models (:pr:`81`)
+* WSClean Spectral Model (:pr:`81`)
 * Compare predict versus MeqTrees (:pr:`79`)
 * Time and channel averaging (:pr:`75`)
 * cupy implementation of `predict_vis` (:pr:`73`)
